pyproject.toml
[tool.poetry]
name = "datascience_projects"
version = "0.1.0"
description = ""
authors = ["smrenato <renatosmx@gmail.com>"]
license = ""
packages = [{include = "src"}]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ipython = "^8.8.0"
rich = "^13.0.1"
typer = "^0.7.0"
dynaconf = "^3.1.11"
mkdocs-material = "^9.0.3"
pymdown-extensions = "^9.9"
dvc = {extras = ["gdrive"], version = "^2.41.1"}
ipykernel = "^6.20.2"
pytest = "^7.2.0"
hydra-core = "^1.3.2"

[tool.poetry.scripts]
dsp = "src.cli:main"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
